Add files via upload
Each question will have a time limit(10s here). If the user is unable to pick the answer within that limit, next question will appear and scored for that question is counted zero.
Time based feature added. The scores now will be generated on the basis of how much of time was left for that particular question. A simple linear function is used. (Ex : If the user finishes one question in 9s, the sore for that question will be 90%).